+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Local `$ref` resolution within fragment-extracted external resources. [#892](https://github.com/Stranger6667/jsonschema/issues/892)
 
+### Removed
+
+- Deprecated `PrimitiveType` & `PrimitiveTypesBitMap`.
+
 ## [0.37.3] - 2025-11-28
 
 ### Fixed